remove all non RecurringSelect files, reorganize structure, update imports
