+++ packages/components/src/RecurringSelect/RecurringSelect.mdx
@@ -1,7 +1,7 @@
 ---
-name: Scheduler
+name: RecurringSelect
 menu: Components
-route: /components/scheduler
+route: /components/recuringselect
 showDirectoryLink: true
 ---
 
@@ -24,7 +24,7 @@ import { useScheduler } from "./useScheduler";
 RecurringSelect, fill me out & rename me please
 
 ```ts
-import { Scheduler } from "@jobber/components/Scheduler";
+import { RecurringSelect } from "@jobber/components/RecurringSelect";
 ```
 
 <Playground>
