--- packages/components/src/Scheduler/ScheduleSummary.css
@@ -1,17 +1,0 @@
-.disabled > h5 {
-  color: var(--color-disabled);
-}
-
-.disabled > p {
-  color: var(--color-disabled);
-}
-
-.disabled dt,
-.disabled dd {
-  color: var(--color-disabled);
-}
-
-.summaryContainer {
-  display: grid;
-  grid-row-gap: var(--space-small);
-}
