--- packages/components/src/Scheduler/ScheduleSummary.css.d.ts
@@ -1,6 +1,0 @@
-declare const styles: {
-  readonly "disabled": string;
-  readonly "summaryContainer": string;
-};
-export = styles;
-
