--- packages/components/src/Scheduler/Scheduler.css.d.ts
@@ -1,8 +1,0 @@
-declare const styles: {
-  readonly "title": string;
-  readonly "datePicker": string;
-  readonly "container": string;
-  readonly "disabled": string;
-};
-export = styles;
-
