--- packages/components/src/Scheduler/index.tsx
@@ -1,1 +1,0 @@
-export { Scheduler } from "./Scheduler";
