--- packages/components/src/Scheduler/scheduleToIceCube.ts
@@ -1,6 +1,0 @@
-import { SchedulerState } from "./types";
-
-// eslint-disable-next-line @typescript-eslint/no-unused-vars
-export const scheduleToIceCube = (schedulerState: SchedulerState) => {
-  return "";
-};
